Add the ability to persist streams created through the API
Save the stream definition on the file system (in the first directory
of the directories list) if the :persist option is set to true in the
stream configuration.

The file name is "<stream-name>-api.edn" to avoid conflicts with
existing files.

The file (and so the stream) is also automatically deleted if the
stream if deleted using the API.